Add command in rst mode to insert external links
Bound to "C-c i". I would like it to use the marked region
automatically as description but had trouble with :default-value of
read-string. Also "C-c C-l" would be better (same kbd as in org mode)
but "C-c C-l" is globbered in rst mode and I have yet to find out how
to unbind it.